power:qpnp-charger: extended charge time from 30s to 100s when battery status is near to full
When charger is online for a long time, such as 12 hour, the battery status changes from charging to full and resumes charging.
The reason is that battery is not full at the first full state. So increse charge time from 30s to 100s when battery status is
near to full. This methods will reduce the number of recharging times

BUG:24316771
Signed-off-by: l00228880 <l00228880@notesmail.huawei.com>
Signed-off-by: Francisco Franco <franciscofranco.1990@gmail.com>
Signed-off-by: dd3boh <dade.garberi@gmail.com>
Signed-off-by: Ryan Andri <ryanandri@linuxmail.org>